+++ packages/engine-server/src/workspace/service.ts
@@ -50,6 +50,7 @@ export enum SyncActionStatus {
   NO_CHANGES = "it has no changes",
   UNCOMMITTED_CHANGES = "it has uncommitted changes",
   NO_REMOTE = "it has no remote",
+  NO_UPSTREAM = "the current branch has no upstream",
   SKIP_CONFIG = "it is configured so",
   NOT_PERMITTED = "user is not permitted to push to one or more vaults",
   NEW = "newly clond repository",
@@ -77,6 +78,17 @@ export type WorkspaceServiceOpts = {
 };
 
 type UrlTransformerFunc = (url: string) => string;
+
+type AddRemoveCommonOpts = {
+  /**
+   * Default: true
+   */
+  updateConfig?: boolean;
+  /**
+   * Default: false
+   */
+  updateWorkspace?: boolean;
+};
 
 export class WorkspaceService {
   public logger: DLogger;
@@ -134,6 +146,7 @@ export class WorkspaceService {
   async addWorkspace({ workspace }: { workspace: DWorkspace }) {
     const allWorkspaces = this.config.workspaces || {};
     allWorkspaces[workspace.name] = _.omit(workspace, ["name", "vaults"]);
+    const config = this.config;
     // update vault
     const newVaults = await _.reduce(
       workspace.vaults,
@@ -141,14 +154,15 @@ export class WorkspaceService {
         const out = await acc;
         out.push(
           await this.addVault({
+            config,
             vault: { ...vault, workspace: workspace.name },
+            updateConfig: false,
           })
         );
         return out;
       },
       Promise.resolve([] as DVault[])
     );
-    const config = this.config;
     config.workspaces = allWorkspaces;
     this.setConfig(config);
     return { vaults: newVaults };
@@ -163,16 +177,16 @@ export class WorkspaceService {
    * @param opts.addToWorkspace - default: false, add to dendron.code-workspace
    * @returns
    */
-  async addVault(opts: {
-    vault: DVault;
-    config?: DendronConfig;
-    writeConfig?: boolean;
-    addToWorkspace?: boolean;
-  }) {
-    const { vault, config, writeConfig, addToWorkspace } = _.defaults(opts, {
+  async addVault(
+    opts: {
+      vault: DVault;
+      config?: DendronConfig;
+    } & AddRemoveCommonOpts
+  ) {
+    const { vault, config, updateConfig, updateWorkspace } = _.defaults(opts, {
       config: this.config,
-      writeConfig: true,
-      addToWorkspace: false,
+      updateConfig: true,
+      updateWorkspace: false,
     });
     config.vaults.unshift(vault);
     // update dup note behavior
@@ -184,10 +198,10 @@ export class WorkspaceService {
     } else if (_.isArray(config.site.duplicateNoteBehavior.payload)) {
       config.site.duplicateNoteBehavior.payload.push(VaultUtils.getName(vault));
     }
-    if (writeConfig) {
+    if (updateConfig) {
       await this.setConfig(config);
     }
-    if (addToWorkspace) {
+    if (updateWorkspace) {
       const wsPath = path.join(this.wsRoot, DENDRON_WS_NAME);
       let out = (await readJSONWithComments(wsPath)) as WorkspaceSettings;
       if (
@@ -342,13 +356,19 @@ export class WorkspaceService {
   }
 
   /**
-   * Remove vaults. Currently doesn't delete any files
+   * Remove vaults. Currently doesn't delete any files.
    * @param param0
    */
-  async removeVault({ vault }: { vault: DVault }) {
+  async removeVault(opts: { vault: DVault } & AddRemoveCommonOpts) {
     const config = this.config;
+    const { vault, updateConfig, updateWorkspace } = _.defaults(opts, {
+      updateConfig: true,
+      updateWorkspace: false,
+    });
     config.vaults = _.reject(config.vaults, (ent) => {
-      const checks = [ent.fsPath === vault.fsPath];
+      const checks = [
+        VaultUtils.getRelPath(ent) === VaultUtils.getRelPath(vault),
+      ];
       if (vault.workspace) {
         checks.push(ent.workspace === vault.workspace);
       }
@@ -377,7 +397,19 @@ export class WorkspaceService {
         );
       }
     }
-    await this.setConfig(config);
+    if (updateConfig) {
+      await this.setConfig(config);
+    }
+    if (updateWorkspace) {
+      const wsPath = path.join(this.wsRoot, DENDRON_WS_NAME);
+      let settings = (await readJSONWithComments(wsPath)) as WorkspaceSettings;
+      const folders = _.reject(
+        settings.folders,
+        (ent) => ent.path === VaultUtils.getRelPath(vault)
+      );
+      settings = assignJSONWithComment({ folders }, settings);
+      writeJSONWithComments(wsPath, settings);
+    }
   }
 
   createConfig() {
@@ -413,10 +445,13 @@ export class WorkspaceService {
     if (opts.createCodeWorkspace) {
       WorkspaceConfig.write(wsRoot, vaults);
     }
-    await Promise.all(
-      vaults.map(async (vault) => {
+    await _.reduce(
+      vaults,
+      async (prev, vault) => {
+        await prev;
         return ws.createVault({ vault });
-      })
+      },
+      Promise.resolve()
     );
     return ws;
   }
@@ -566,6 +601,8 @@ export class WorkspaceService {
           // It's impossible to pull if there is no remote, or if there are tracked files that have changes
           if (!(await git.hasRemote()))
             return { repo, vaults, status: SyncActionStatus.NO_REMOTE };
+          if (_.isUndefined(await git.getUpstream()))
+            return { repo, vaults, status: SyncActionStatus.NO_UPSTREAM };
           if (!(await this.shouldVaultsSync("pull", rootVaults)))
             return { repo, vaults, status: SyncActionStatus.SKIP_CONFIG };
           if (await git.hasChanges({ untrackedFiles: "no" }))
@@ -599,8 +636,20 @@ export class WorkspaceService {
         async (rootVaults: [string, DVault[]]): Promise<SyncActionResult> => {
           const [repo, vaults] = rootVaults;
           const git = new Git({ localUrl: repo });
+
           if (!(await git.hasRemote()))
             return { repo, vaults, status: SyncActionStatus.NO_REMOTE };
+          const upstream = await git.getUpstream();
+          if (_.isUndefined(upstream))
+            return { repo, vaults, status: SyncActionStatus.NO_UPSTREAM };
+          if (
+            (await git.diff({
+              nameOnly: true,
+              oldCommit: upstream,
+              newCommit: "HEAD",
+            })) === ""
+          )
+            return { repo, vaults, status: SyncActionStatus.NO_CHANGES };
           if (!(await this.shouldVaultsSync("push", rootVaults)))
             return { repo, vaults, status: SyncActionStatus.SKIP_CONFIG };
           if (!_.every(_.map(vaults, this.user.canPushVault)))
@@ -693,7 +742,6 @@ export class WorkspaceService {
           }
           const spath = await this.seedService.cloneSeed({
             seed: resp,
-            wsRoot,
           });
           seedResults.push({
             id,