+++ packages/engine-test-utils/src/__tests__/engine-server/enginev2.spec.ts
@@ -136,6 +136,7 @@ describe("engine, cache", () => {
 describe("engine, notes/", () => {
   const nodeType = "NOTES";
 
+  // EXAMPLE of running a single test
   // test.only("bond", async () => {
   //   const preset = getPreset({key: "NOTE_REF", nodeType: "NOTES", presetName: "rename", presets: ENGINE_PRESETS})
   //   const { testFunc, ...opts } = preset;